fix(alerts): restore 432 grouping classification
Prefer auth_token_logs HTTP status for usage-limit grouping and keep request_logs as fallback.

Split oversized alerts and runtime logging tests into focused modules.

Raise the models budget exception minimally for the new DTO surface.
