+++ src/server/tests.rs
@@ -32,6 +32,7 @@ mod tests {
     mod admin_token_filters_and_maintenance;
     mod admin_users_and_tokens;
     mod alerts_and_ha;
+    mod alerts_and_ha_dashboard_defaults;
     mod alerts_and_ha_event_exports;
     mod alerts_and_ha_node_state;
     mod alerts_and_ha_sync_recovery;
