+++ tests/rust_source_line_budgets.rs
@@ -6,7 +6,7 @@ const IGNORE_DIRS: &[&str] = &["target", ".git"];
 const EXCEPTIONS: &[(&str, usize, &str)] = &[
     (
         "src/models.rs",
-        3150,
+        3200,
         "Alert center semantic DTO growth and the new admin analysis pressure snapshot contracts both extend the shared model surface while the broader extraction pass remains pending.",
     ),
     (
